% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewMouseSciaticNerve.R
\name{viewMouseSciaticNerve}
\alias{viewMouseSciaticNerve}
\title{View MouseSciaticNerve data in the scClustViz Shiny app}
\usage{
viewMouseSciaticNerve(dataset, outPath = "./")
}
\arguments{
\item{dataset}{One of "Inj9dBeads", "Inj9dBeadsMesenchymal", "Inj9dFACS", "InjCombined", "InjUninjMesenchymalCombined", "UninjMesenchymal".
Referring to the different datasets in this package, which each represent single-cell RNAseq data collected from different Injured/Uninjured Sciatic Nerve.}

\item{outPath}{Default = "./" (the working directory). Specify the directory
used to save/load any analysis files you generate while exploring the
\code{MouseSciaticNerve} data.}
}
\value{
The function causes the scClustViz Shiny GUI app to open in a
  seperate window.
}
\description{
A wrapper function to view one of the six \code{MouseSciaticNerve} datasets / analyses in the
\code{scClustViz} Shiny app.
}
\examples{
  viewMouseSciaticNerve("Inj9dBeads")

}
\seealso{
\url{https://baderlab.github.io/scClustViz} for information on
  \code{scClustViz}.
}
